django/trello/accounts/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Cerberus==1.3.4
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==3.2
django-cors-headers==3.12.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
jsonfield==3.1.0
pkg_resources==0.0.0
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
